break out more functionality into separate source files
