+++ include/clock_mode.h
@@ -1,0 +1,6 @@
+#ifndef __CLOCK_MODE_H__
+#define __CLOCK_MODE_H__
+
+extern void clock_mode();
+
+#endif
