+++ include/idle_mode.h
@@ -1,0 +1,6 @@
+#ifndef __IDLE_MODE_H__
+#define __IDLE_MODE_H__
+
+extern void idle_mode();
+
+#endif
