+++ include/options_mode.h
@@ -1,0 +1,7 @@
+#ifndef __OPTIONS_MODE_H__
+#define __OPTIONS_MODE_H__
+
+extern void options_mode();
+extern void options_menu();
+
+#endif
