+++ include/sleep_mode.h
@@ -1,0 +1,6 @@
+#ifndef __SLEEP_H__
+#define __SLEEP_H__
+
+extern void sleep_mode();
+
+#endif
