+++ include/test_mode.h
@@ -1,0 +1,8 @@
+#ifndef __TEST_MODE_H__
+#define __TEST_MODE_H__
+
+#define TEST_TIMEOUT 1000
+
+extern void test_mode();
+
+#endif
