+++ include/timer_mode.h
@@ -1,0 +1,6 @@
+#ifndef __TIMER_MODE_H__
+#define __TIMER_MODE_H__
+
+extern void timer_mode();
+
+#endif
