+++ include/tools_mode.h
@@ -1,0 +1,6 @@
+#ifndef __TOOLS_MODE_H__
+#define __TOOLS_MODE_H__
+
+extern void tools_menu();
+
+#endif
