+++ src/billboard.cpp
@@ -1,4 +1,4 @@
-#include <HT16K33Disp.h>
+// #include <HT16K33Disp.h>
 #include "billboard.h"
 
 Billboard::Billboard(char * buffer, int show_delay, int frame_delay){
