+++ src/billboards_handler.cpp
@@ -1,5 +1,5 @@
 #include <Arduino.h>
-#include <HT16K33Disp.h>
+// #include <HT16K33Disp.h>
 #include "billboard.h"
 #include "billboards_handler.h"
