+++ src/clock_mode.cpp
@@ -1,0 +1,5 @@
+#include "prompts.h"
+
+void clock_mode(){
+  clock_prompt(0, 0, 12);
+}
