+++ src/timer_mode.cpp
@@ -1,0 +1,5 @@
+#include "prompts.h"
+
+void timer_mode(){
+  timer_prompt(0, 0, 0);
+}
